🗃Refactor: 프레젠테이션 컴포넌트 수정
탭 별 화면을 분리한다.
